Streamlit/Heroku Deployment/requirements.txt
streamlit
pandas
numpy
Pillow
sklearn